Fix reload property
